Reduce turbo repeat rate
The TASCAM DR-40 cannot always keep up with a repeat byte sent at 40
times per second, especially with the mic input level button, but it can
(usually) keep up with 20 times per second.

I may remove the "turbo" feature in the future as it doesn't seem to be
as useful as I thought it might be.